Initial Flame integration
