+++ lib/main.dart
@@ -1,179 +1,123 @@
+import 'package:flame/game.dart';
+import 'package:flame/components.dart';
+import 'package:flame/input.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter/foundation.dart';
-import 'dart:async';
-import 'package:shake/shake.dart';
 
-void main() => runApp(const ChessBoardApp());
-
-class ChessBoardApp extends StatelessWidget {
-  const ChessBoardApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const MaterialApp(
-      home: ChessBoardScreen(),
-    );
-  }
+void main() {
+  runApp(
+    GameWidget(
+      game: ChessBoardGame(),
+    ),
+  );
 }
 
-class ChessBoardScreen extends StatefulWidget {
-  const ChessBoardScreen({super.key});
-
-  @override
-  ChessBoardScreenState createState() => ChessBoardScreenState();
-}
-
-class ChessBoardScreenState extends State<ChessBoardScreen> {
-  late int boardSize;
-  late double cellSize;
+class ChessBoardGame extends FlameGame with TapDetector, PanDetector {
+  static const int boardSize = 30;
+  late final SpriteComponent background;
+  late final PositionComponent boardComponent;
   int userRow = 0;
   int userCol = 0;
-  final FocusNode _focusNode = FocusNode();
   Set<String> highlightedCells = {};
-  Timer? _movementTimer;
   bool _blinkGreen = false;
 
   @override
-  void initState() {
-    super.initState();
-    _focusNode.requestFocus();
+  Future<void> onLoad() async {
+    await super.onLoad();
 
-    if (!kIsWeb) {
-      ShakeDetector.autoStart(
-        onPhoneShake: () {
-          clearEverything();
-        },
-      );
+    // Load background
+    background = SpriteComponent(
+      sprite: await loadSprite('background.png'),
+      size: size,
+    );
+    add(background);
+
+    // Create board
+    boardComponent = PositionComponent();
+    add(boardComponent);
+    createBoard();
+  }
+
+  void createBoard() {
+    double cellSize = size.x / boardSize;
+    for (int row = 0; row < boardSize; row++) {
+      for (int col = 0; col < boardSize; col++) {
+        bool isWhiteCell = (row + col) % 2 == 0;
+        RectangleComponent cell = RectangleComponent(
+          position: Vector2(col * cellSize, row * cellSize),
+          size: Vector2.all(cellSize),
+          paint: Paint()..color = isWhiteCell ? Colors.white : Colors.black,
+        );
+        boardComponent.add(cell);
+      }
+    }
+    updateUserPosition();
+  }
+
+  void updateUserPosition() {
+    double cellSize = size.x / boardSize;
+    RectangleComponent userCell = RectangleComponent(
+      position: Vector2(userCol * cellSize, userRow * cellSize),
+      size: Vector2.all(cellSize),
+      paint: Paint()..color = Colors.green,
+    );
+    boardComponent.add(userCell);
+  }
+
+  @override
+  void onTapDown(TapDownInfo info) {
+    Vector2 tapPosition = info.eventPosition.game;
+    double cellSize = size.x / boardSize;
+    int tappedRow = (tapPosition.y / cellSize).floor();
+    int tappedCol = (tapPosition.x / cellSize).floor();
+
+    if ((tappedRow == userRow && (tappedCol == userCol - 1 || tappedCol == userCol + 1)) ||
+        (tappedCol == userCol && (tappedRow == userRow - 1 || tappedRow == userRow + 1))) {
+      userRow = tappedRow;
+      userCol = tappedCol;
+      updateHighlightedCells();
+      updateUserPosition();
     }
   }
 
   @override
-  void dispose() {
-    _focusNode.dispose();
-    _movementTimer?.cancel();
-    super.dispose();
-  }
-
-  void calculateBoardSize(BuildContext context) {
-    final screenSize = MediaQuery.of(context).size;
-    final shortestSide = screenSize.shortestSide;
-
-    const desiredCellSize = 30.0;
-    const minBoardSize = 10;
-    const maxBoardSize = 50;
-
-    boardSize = (shortestSide / desiredCellSize).floor();
-    boardSize = boardSize.clamp(minBoardSize, maxBoardSize);
-
-    cellSize = shortestSide / boardSize;
-
-    userRow = userRow.clamp(0, boardSize - 1);
-    userCol = userCol.clamp(0, boardSize - 1);
-  }
-
-  Widget buildBoard() {
-    return SingleChildScrollView(
-      child: SingleChildScrollView(
-        scrollDirection: Axis.horizontal,
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: List.generate(boardSize, (row) {
-            return Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: List.generate(boardSize, (col) {
-                bool isUserBlock = row == userRow && col == userCol;
-                bool isWhiteCell = (row + col) % 2 == 0;
-                bool isHighlighted = highlightedCells.contains('$row,$col');
-                return GestureDetector(
-                  onTap: () => handleTap(row, col),
-                  child: Container(
-                    width: cellSize,
-                    height: cellSize,
-                    color: isUserBlock
-                        ? Colors.green
-                        : _blinkGreen && isWhiteCell
-                        ? Colors.green
-                        : isHighlighted
-                        ? Colors.purple
-                        : isWhiteCell
-                        ? Colors.white
-                        : Colors.black,
-                    child: Center(
-                      child: Text(
-                        isUserBlock ? 'U' : '',
-                        style: TextStyle(
-                            color: isUserBlock ? Colors.white : Colors.black,
-                            fontSize: cellSize / 2),
-                      ),
-                    ),
-                  ),
-                );
-              }),
-            );
-          }),
-        ),
-      ),
-    );
-  }
-
-  void handleTap(int row, int col) {
-    if ((row == userRow && (col == userCol - 1 || col == userCol + 1)) ||
-        (col == userCol && (row == userRow - 1 || row == userRow + 1))) {
-      setState(() {
-        userRow = row;
-        userCol = col;
-        updateHighlightedCells();
-      });
-    }
-  }
-
-  void handleKeyboard(KeyEvent event) {
-    if (event is KeyDownEvent) {
-      if (event.logicalKey == LogicalKeyboardKey.space) {
-        clearEverything();
+  void onPanUpdate(DragUpdateInfo info) {
+    const int sensitivity = 5;
+    if (info.delta.game.length > sensitivity) {
+      if (info.delta.game.x.abs() > info.delta.game.y.abs()) {
+        // Horizontal swipe
+        if (info.delta.game.x > 0) {
+          move('Right');
+        } else {
+          move('Left');
+        }
       } else {
-        startMovement(event.logicalKey.keyLabel);
+        // Vertical swipe
+        if (info.delta.game.y > 0) {
+          move('Down');
+        } else {
+          move('Up');
+        }
       }
-    } else if (event is KeyUpEvent) {
-      stopMovement();
     }
-  }
-
-  void startMovement(String keyLabel) {
-    move(keyLabel);
-    _movementTimer?.cancel();
-    _movementTimer = Timer.periodic(const Duration(milliseconds: 100), (timer) {
-      move(keyLabel);
-    });
-  }
-
-  void stopMovement() {
-    _movementTimer?.cancel();
   }
 
   void move(String direction) {
-    setState(() {
-      switch (direction) {
-        case 'Arrow Up':
-        case 'Up':
-          if (userRow > 0) userRow--;
-          break;
-        case 'Arrow Down':
-        case 'Down':
-          if (userRow < boardSize - 1) userRow++;
-          break;
-        case 'Arrow Left':
-        case 'Left':
-          if (userCol > 0) userCol--;
-          break;
-        case 'Arrow Right':
-        case 'Right':
-          if (userCol < boardSize - 1) userCol++;
-          break;
-      }
-      updateHighlightedCells();
-    });
+    switch (direction) {
+      case 'Up':
+        if (userRow > 0) userRow--;
+        break;
+      case 'Down':
+        if (userRow < boardSize - 1) userRow++;
+        break;
+      case 'Left':
+        if (userCol > 0) userCol--;
+        break;
+      case 'Right':
+        if (userCol < boardSize - 1) userCol++;
+        break;
+    }
+    updateHighlightedCells();
+    updateUserPosition();
   }
 
   void updateHighlightedCells() {
@@ -191,75 +135,14 @@ class ChessBoardScreenState extends State<ChessBoardScreen> {
         highlightedCells.add(cell);
       }
     }
+    // Update highlighted cells visually
+    // This part needs to be implemented
   }
 
   void clearEverything() {
-    setState(() {
-      _blinkGreen = true;
-      highlightedCells.clear();
-    });
-
-    Future.delayed(const Duration(milliseconds: 500), () {
-      setState(() {
-        _blinkGreen = false;
-      });
-    });
-  }
-
-  void handleSwipe(DragUpdateDetails details) {
-    const int sensitivity = 5;
-    if (details.delta.dx.abs() > sensitivity || details.delta.dy.abs() > sensitivity) {
-      if (details.delta.dx.abs() > details.delta.dy.abs()) {
-        // Horizontal swipe
-        if (details.delta.dx > 0) {
-          move('Right');
-        } else {
-          move('Left');
-        }
-      } else {
-        // Vertical swipe
-        if (details.delta.dy > 0) {
-          move('Down');
-        } else {
-          move('Up');
-        }
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    calculateBoardSize(context);
-    return Scaffold(
-      appBar: AppBar(title: Text('${boardSize}x$boardSize Grid Movement')),
-      body: Column(
-        children: [
-          Expanded(
-            child: KeyboardListener(
-              focusNode: _focusNode,
-              onKeyEvent: (KeyEvent event) {
-                handleKeyboard(event);
-              },
-              child: GestureDetector(
-                onPanUpdate: (details) {
-                  handleSwipe(details);
-                },
-                child: Center(
-                  child: buildBoard(),
-                ),
-              ),
-            ),
-          ),
-          if (kIsWeb)
-            Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: ElevatedButton(
-                onPressed: clearEverything,
-                child: const Text('Clear Everything'),
-              ),
-            ),
-        ],
-      ),
-    );
+    _blinkGreen = true;
+    highlightedCells.clear();
+    // Implement blinking effect
+    // This part needs to be implemented
   }
 }
